requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
google-genai>=1.0.0
SQLAlchemy>=2.0.0
psycopg[binary]>=3.1.0
requests>=2.31.0
google-cloud-storage>=2.14.0
